+++ libs/docs/ui5-webcomponents/form/examples/empty-span.ts
@@ -1,5 +1,6 @@
 import { NgStyle } from '@angular/common';
 import { Component, signal } from '@angular/core';
+import { UI5WrapperCustomEvent } from '@fundamental-ngx/ui5-webcomponents-base';
 import { Form } from '@fundamental-ngx/ui5-webcomponents/form';
 import { FormItem } from '@fundamental-ngx/ui5-webcomponents/form-item';
 import { Input as UI5Input } from '@fundamental-ngx/ui5-webcomponents/input';
@@ -31,7 +32,7 @@ export class FormEmptySpanSample {
     readonly containerWidth = signal('1250px');
     readonly pageSize = signal('L');
 
-    onSliderChange(event: CustomEvent): void {
+    onSliderChange(event: UI5WrapperCustomEvent<Slider, 'ui5Change'>): void {
         const width = (event.target?.['value'] / 100) * 1500;
         this.containerWidth.set(`${width}px`);
         this.pageSize.set(this.getLayoutByWidth(width));