+++ libs/docs/ui5-webcomponents/list/examples/drag-and-drop.ts
@@ -1,10 +1,8 @@
 import { Component, signal } from '@angular/core';
+import type { UI5WrapperCustomEvent } from '@fundamental-ngx/ui5-webcomponents-base';
 import { MovePlacement } from '@fundamental-ngx/ui5-webcomponents-base/types';
 import { List } from '@fundamental-ngx/ui5-webcomponents/list';
 import { ListItemStandard } from '@fundamental-ngx/ui5-webcomponents/list-item-standard';
-
-import { UI5CustomEvent } from '@ui5/webcomponents-base';
-import { default as _List } from '@ui5/webcomponents/dist/List.js';
 
 // Import the icon used in this example
 import '@ui5/webcomponents-icons/dist/checklist-item.js';
@@ -18,7 +16,7 @@ import '@ui5/webcomponents-icons/dist/checklist-item.js';
 export class ListDragAndDropExample {
     readonly items = signal(['Item #1', 'Item #2', 'Item #3', 'Item #4']);
 
-    onItemMove(event: any): void {
+    onItemMove(event: UI5WrapperCustomEvent<List, 'ui5Move'>): void {
         const { destination, source } = event.detail;
 
         switch (destination.placement) {
@@ -34,7 +32,7 @@ export class ListDragAndDropExample {
         }
     }
 
-    onItemMoveOver(event: UI5CustomEvent<_List, 'move-over'>): void {
+    onItemMoveOver(event: UI5WrapperCustomEvent<List, 'ui5MoveOver'>): void {
         const { source } = event.detail;
 
         if (!(event.target as HTMLElement).contains(source.element)) {
@@ -44,7 +42,7 @@ export class ListDragAndDropExample {
         this.handleBeforeItemMove(event);
     }
 
-    private handleBeforeItemMove(event: UI5CustomEvent<_List, 'move-over'>): void {
+    private handleBeforeItemMove(event: UI5WrapperCustomEvent<List, 'ui5MoveOver'>): void {
         const { destination } = event.detail;
 
         if (destination.placement === 'Before' || destination.placement === 'After') {