+++ CHANGELOG.md
@@ -3,6 +3,71 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.58.0-rc.25](https://github.com/SAP/fundamental-ngx/compare/v0.58.0-rc.24...v0.58.0-rc.25) (2025-11-26)
+
+
+### Bug Fixes
+
+* **core:** add correct icons for form messages ([#13616](https://github.com/SAP/fundamental-ngx/issues/13616)) ([c9eb507](https://github.com/SAP/fundamental-ngx/commit/c9eb507d1c8806229881a03ed2c6fc37d841c940))
+
+
+
+
+
+# [0.58.0-rc.24](https://github.com/SAP/fundamental-ngx/compare/v0.58.0-rc.23...v0.58.0-rc.24) (2025-11-26)
+
+
+### Bug Fixes
+
+* **core:** add fix for Combobox incorrect behavior ([#13613](https://github.com/SAP/fundamental-ngx/issues/13613)) ([f62b7e1](https://github.com/SAP/fundamental-ngx/commit/f62b7e1ba5f0abe8d88cd74b8c7b252da349189d))
+
+
+### Features
+
+* add types for wrapper event emitters ([#13578](https://github.com/SAP/fundamental-ngx/issues/13578)) ([73a0431](https://github.com/SAP/fundamental-ngx/commit/73a04318958de80780c5034517cedcf2902e5e50))
+
+
+
+
+
+# [0.58.0-rc.23](https://github.com/SAP/fundamental-ngx/compare/v0.58.0-rc.22...v0.58.0-rc.23) (2025-11-26)
+
+
+### Bug Fixes
+
+* **platform:** show navigation indicator on Settings Generator items on small screens ([#13612](https://github.com/SAP/fundamental-ngx/issues/13612)) ([8cd8c5e](https://github.com/SAP/fundamental-ngx/commit/8cd8c5edfb23ef5bb68b32615302878c0b5f8938))
+
+
+
+
+
+# [0.58.0-rc.22](https://github.com/SAP/fundamental-ngx/compare/v0.58.0-rc.21...v0.58.0-rc.22) (2025-11-26)
+
+
+### Bug Fixes
+
+* **core:** add correct font family for table header text ([#13611](https://github.com/SAP/fundamental-ngx/issues/13611)) ([7b56817](https://github.com/SAP/fundamental-ngx/commit/7b5681786c1863e5f32eaa736a4c7d7a7b4d5053))
+
+
+
+
+
+# [0.58.0-rc.21](https://github.com/SAP/fundamental-ngx/compare/v0.58.0-rc.20...v0.58.0-rc.21) (2025-11-26)
+
+**Note:** Version bump only for package fundamental-ngx
+
+
+
+
+
+# [0.58.0-rc.20](https://github.com/SAP/fundamental-ngx/compare/v0.58.0-rc.19...v0.58.0-rc.20) (2025-11-26)
+
+**Note:** Version bump only for package fundamental-ngx
+
+
+
+
+
 # [0.58.0-rc.19](https://github.com/SAP/fundamental-ngx/compare/v0.58.0-rc.18...v0.58.0-rc.19) (2025-11-25)
 
 **Note:** Version bump only for package fundamental-ngx